+++ src/main/java/fr/unice/miage/sd/tinydfs/nodes/Slave.java
@@ -20,6 +20,6 @@ public interface Slave extends Remote {
 
 	public List<byte[]> subRetrieve(String filename) throws RemoteException;
 	
-	public float sizeOf(String filename) throws RemoteException;
+	public long sizeOf(String filename) throws RemoteException;
 
 }
